работа с бд

+++ src/main/java/com/webshop/simplewebapplication/controller/DeleteController.java
@@ -1,9 +1,13 @@
 package com.webshop.simplewebapplication.controller;
 
+import com.webshop.simplewebapplication.Service.CartService;
 import com.webshop.simplewebapplication.Service.ItemService;
 import com.webshop.simplewebapplication.model.Item;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
@@ -13,13 +17,29 @@ import java.util.List;
 @RestController
 public class DeleteController {
 
+    @Autowired
+    ItemService itemService;
+
+    @Autowired
+    CartService cartService;
+
     static final Logger logger = LoggerFactory.getLogger(DeleteController.class);
 
     @RequestMapping(path = "/item/{id}/delete", method = RequestMethod.POST)
     public ModelAndView delete(@PathVariable("id") int id) {
-        ItemService itemService = new ItemService();
-        itemService.deleteItem(id);
-        logger.info("Deleted item with item_id: " + id);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserName = authentication.getName();
+
+        Item item = itemService.findById(id);
+
+
+        if (item.getUserName().equals(currentUserName)){
+            itemService.deleteItem(id);
+            logger.info("Deleted item with item_id: " + id);
+        }
+        else {
+            logger.info("Deleted item with item_id: " + id + "FAILED. Not enough rules");
+        }
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.setViewName("index");
         return modelAndView;
@@ -27,7 +47,6 @@ public class DeleteController {
 
     @RequestMapping(path = "/item/{id}/delete/cart", method = RequestMethod.POST)
     public ModelAndView deleteFromCart(@PathVariable("id") int id) {
-        ItemService itemService = new ItemService();
         itemService.deleteFromCart(id, false);
         logger.info("Deleted item from cart with item_id: " + id);
         ModelAndView modelAndView = new ModelAndView();
@@ -37,7 +56,6 @@ public class DeleteController {
 
     @RequestMapping(path = "/items/deleteAll/cart", method = RequestMethod.POST)
     public ModelAndView deleteAllFromCart() {
-        ItemService itemService = new ItemService();
         List<Item> items = itemService.findAllInCart();
         for (Item item:items){
             itemService.deleteFromCart(item.getId(), true);
